+++ resdk/resources/__init__.py
@@ -3,5 +3,6 @@
 from .collection import Collection
 from .data import Data
 from .sample import Sample
+from .process import Process
 
-__all__ = ['Collection', 'Data', 'Sample']
+__all__ = ['Collection', 'Data', 'Sample', 'Process']